\name{StatComp21048-package}
\alias{StatComp21048-package}
\alias{StatComp21048}
\docType{package}
\title{
  Final homework of statistical computations
}
\description{
  It includes the homework of the whole semester and two functions of my research field.
}
\author{
StatComp21048
Maintainer: StatComp21048 <zheng1003@mail.ustc.edu.cn>
}
\keyword{ package }

